create modules to train,test,analyze predictions of dictionarized dataframe
Create a dictionary for each of X_train, X_test, y_train, y_test that stores the desired split for each variety. Then create a dictionary of classifiers to store a fit random forest for each variety. Then we can test it. The results seem to not be doing as well as expected